Move Numba nlinalg tests to test_nlinalg
